Added dirHandler in 'Utility.h' + improvements
- Fixed some cases where the query string is not desirably splitted
- Fixed warning: comparison between signed and unsigned

+++ include/Base.h
@@ -63,7 +63,7 @@ public:
     // get word that has a hash id
     std::string getWordById (int id) const
     {
-        return  id >= 1 && id <= words.size() ? words[id - 1] : "";
+        return  id >= 1 && id <= (int)words.size() ? words[id - 1] : "";
     }
 };
